+++ KMSCalendar/KMSCalendar/KMSCalendar.csproj
@@ -48,4 +48,8 @@
       <Generator>MSBuild:UpdateDesignTimeXaml</Generator>
     </EmbeddedResource>
   </ItemGroup>
+
+  <ItemGroup>
+    <Folder Include="Services\Data\" />
+  </ItemGroup>
 </Project>
