+++ KMSCalendar/KMSCalendar/Models/Data/User.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace KMSCalendar.Models.Entities
+namespace KMSCalendar.Models.Data
 {
     public class User : TableData
     {
