+++ KMSCalendar/KMSCalendar/Services/IDataStore.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Threading.Tasks;
 
-using KMSCalendar.Models.Entities;
+using KMSCalendar.Models.Data;
 
 namespace KMSCalendar.Services
 {
