+++ KMSCalendar/KMSCalendar/Services/MockDataStore.cs
@@ -2,7 +2,7 @@
 using System.Linq;
 using System.Threading.Tasks;
 
-using KMSCalendar.Models.Entities;
+using KMSCalendar.Models.Data;
 
 namespace KMSCalendar.Services
 {
