+++ KMSCalendar/KMSCalendar/ViewModels/AssignmentDetailViewModel.cs
@@ -3,7 +3,7 @@ using System.Windows.Input;
 
 using Xamarin.Forms;
 
-using KMSCalendar.Models.Entities;
+using KMSCalendar.Models.Data;
 using KMSCalendar.Services;
 
 namespace KMSCalendar.ViewModels
