+++ KMSCalendar/KMSCalendar/ViewModels/AssignmentViewModel.cs
@@ -7,7 +7,7 @@ using System.Windows.Input;
 
 using Xamarin.Forms;
 
-using KMSCalendar.Models.Entities;
+using KMSCalendar.Models.Data;
 using KMSCalendar.Services;
 using KMSCalendar.Views;
