+++ KMSCalendar/KMSCalendar/ViewModels/MainViewModel.cs
@@ -1,0 +1,6 @@
+﻿namespace KMSCalendar.ViewModels
+{
+    public class MainViewModel : BaseViewModel
+    {
+    }
+}
