+++ KMSCalendar/KMSCalendar/ViewModels/MenuViewModel.cs
@@ -1,0 +1,7 @@
+﻿namespace KMSCalendar.ViewModels
+{
+    public class MenuViewModel : BaseViewModel
+    {
+
+    }
+}
