+++ KMSCalendar/KMSCalendar/ViewModels/SignUpViewModel.cs
@@ -6,7 +6,7 @@ using Xamarin.Forms;
 using KMSCalendar.Views;
 using System.Threading.Tasks;
 using KMSCalendar.Models;
-using KMSCalendar.Models.Entities;
+using KMSCalendar.Models.Data;
 using KMSCalendar.Services;
 
 namespace KMSCalendar.ViewModels
