+++ KMSCalendar/KMSCalendar/Views/ClassSearchPage.xaml.cs
@@ -5,7 +5,7 @@ using Xamarin.Forms;
 using Xamarin.Forms.Xaml;
 
 using KMSCalendar.Controls;
-using KMSCalendar.Models.Entities;
+using KMSCalendar.Models.Data;
 using KMSCalendar.ViewModels;
 
 namespace KMSCalendar.Views
